+++ context/app/static/js/components/workspaces/WorkspaceDetails.jsx
@@ -1,8 +1,7 @@
-import React, { useContext } from 'react';
+import React from 'react';
 
-import { AppContext } from 'js/components/Providers';
-import OutboundIconLink from 'js/shared-styles/Links/iconLinks/OutboundIconLink';
-import { condenseJobs, startJob } from './utils';
+import { condenseJobs } from './utils';
+import JobLink from './JobLink';
 
 function WorkspaceDetails({ workspace }) {
   const job = condenseJobs(workspace.jobs);
@@ -13,41 +12,12 @@ function WorkspaceDetails({ workspace }) {
         {workspace.name}
       </JobLink>
       {' | '}
-      <JobDetailsDetails job={job} />
+      <JobStatus job={job} />
     </b>
   );
 }
 
-function JobLink({ workspace, job, children }) {
-  const { workspacesEndpoint, workspacesToken } = useContext(AppContext);
-
-  function createHandleStart(workspaceId) {
-    async function handleStart() {
-      startJob({ workspaceId, workspacesEndpoint, workspacesToken });
-      // TODO: Open new tab
-      // eslint-disable-next-line no-alert
-      alert('TODO: Open a new tab that will poll until the job is started.');
-    }
-    return handleStart;
-  }
-
-  if (job.allowNew) {
-    const handleStart = createHandleStart(workspace.id);
-    return (
-      <OutboundIconLink>
-        <button onClick={handleStart} type="submit" style={{ all: 'unset' }}>
-          {children}
-        </button>
-      </OutboundIconLink>
-    );
-  }
-  if (job?.url) {
-    return <OutboundIconLink href={job.url}>{children}</OutboundIconLink>;
-  }
-  return children;
-}
-
-function JobDetailsDetails({ job }) {
+function JobStatus({ job }) {
   if (!job.status) {
     return null;
   }
